Remove unnecessary OTLibraryAccessor, expand BookAccessor
