+++ include/eeprom/core/types.hpp
@@ -7,7 +7,7 @@ namespace types {
 using address = uint16_t;
 
 // 0-64
-using book_address = uint8_t;
+// using book_address = uint8_t;
 
 // 0-8
 using data_length = uint16_t;
